// apps/admin/src/generated/api/schemas/departmentOptionDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentOptionDtoStatus } from "./departmentOptionDtoStatus";

export interface DepartmentOptionDto {
  id: string;
  code: string;
  name: string;
  /** @nullable */
  parentId?: string | null;
  status: DepartmentOptionDtoStatus;
}